feat(API): Adiciona validação de sessão e role nas rotas admin

+++ src/app/api/admin/users/[id]/route.ts
@@ -1,6 +1,8 @@
 import bcrypt from "bcryptjs";
 import { NextRequest, NextResponse } from "next/server";
+import { getServerSession } from "next-auth";
 
+import { authOptions } from "@/lib/auth";
 import { deleteUser, getUserById, updateUser } from "@/services/users";
 import { handleDatabaseError } from "@/utils/handleDbError";
 
@@ -8,6 +10,14 @@ export async function GET(
   _: NextRequest,
   { params }: { params: { id: string } },
 ) {
+  const session = await getServerSession(authOptions);
+  if (!session || session.user?.role !== "admin") {
+    return NextResponse.json(
+      { error: true, message: "Acesso negado." },
+      { status: 401 },
+    );
+  }
+
   try {
     if (!params.id)
       return NextResponse.json(
@@ -26,6 +36,14 @@ export async function DELETE(
   _: NextRequest,
   { params }: { params: { id: string } },
 ) {
+  const session = await getServerSession(authOptions);
+  if (!session || session.user?.role !== "admin") {
+    return NextResponse.json(
+      { error: true, message: "Acesso negado." },
+      { status: 401 },
+    );
+  }
+
   try {
     const deleted = await deleteUser(Number(params.id));
     if (!deleted)
@@ -42,6 +60,14 @@ export async function PUT(
   req: NextRequest,
   { params }: { params: { id: string } },
 ) {
+  const session = await getServerSession(authOptions);
+  if (!session || session.user?.role !== "admin") {
+    return NextResponse.json(
+      { error: true, message: "Acesso negado." },
+      { status: 401 },
+    );
+  }
+
   try {
     const body = await req.json();
